test: keep recordings kind contract helpers out of runtime coverage
