+++ pkg/services/recordings/internal/events/kinds/openapi_mapping_contract_test.go
